requirements.txt
flask>=2.0.0
python-dotenv>=0.19.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-cerebras>=0.0.1
langchain-tavily>=0.1.0
pyswisseph>=2.0.0
pytz>=2021.1
timezonefinder>=5.0.0
geopy>=2.0.0
requests>=2.25.0
tavily-python>=0.1.0
python-dateutil>=2.8.0
pydantic>=2.0.0 